Refactored the northbound/e2/manager.go in to control.go and telemetry.go
