+++ go.mod
@@ -4,14 +4,11 @@ go 1.13
 
 require (
 	github.com/gogo/protobuf v1.3.1
-	github.com/google/go-cmp v0.3.0 // indirect
-	github.com/google/uuid v1.1.1 // indirect
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/pkg/errors v0.8.1 // indirect
-	github.com/prometheus/common v0.4.1 // indirect
+	github.com/onosproject/onos-lib-go v0.0.0-20200218172905-db65cc8e1049
+	github.com/onosproject/onos-ran v0.0.0-20200218171051-a85c1a67d9f8 // indirect
+	github.com/sergi/go-diff v1.1.0 // indirect
 	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
-	google.golang.org/grpc v1.26.0
+	google.golang.org/grpc v1.27.1
 	googlemaps.github.io/maps v0.0.0-20200124220646-5b7f2815585f
 	gotest.tools v2.2.0+incompatible
-	k8s.io/klog v1.0.0
 )
